Fixed the JAR metadata in build.gradle and improved test coverage.
The build script unfortunately was referencing incorrect paths from
prior to the code reorganisation to restrict visibility, this has now
been corrected.

A few tests have been added to exercise more of the code to get the
coverage percentage up - because those badges are basically a game and
you know you want to get it up to 100% at some point.