Implement StartupConfig method for checking whether to expose a class.
It is now possible to implement a method that is passed the
fully-qualified class string and returns true if that class should be
modified to have itself, its methods and its fields made public and,
where possible, non-final.

Additionally, documentation has been added for all packages to improve
the JavaDoc.