Unify TypePool usage and fix up class exposure.
There was an issue where nested classes did not have their visibility
set due to having extra attributes. This then subsequently led to the
discovery of an issue where all types need to be resolveable in order
for the visitor to be able to apply the fixes.

Additionally the JVM passes class names with the foo/bar format, so we
convert to dot notation before running transformation.

All public-facing APIs remain the same.