Update ByteBuddy dependency, Add a test for before=true constructor
We are now using the latest ByteBuddy release and consequently, the
workaround in ClassExposer has been deleted. Additionally, a test for
before=true in constructor interception has been added and behaviour has
been corrected to make the test pass.

This was specifically a problem where, if you are intercepting before
the underlying constructor is called, you cannot obtain the "this"
parameter because it is uninitialised and the JVM will not allow it.

As a result, a "before" constructor interception can only access the
arguments that will later reach the constructor.